kernel: disable usage of fb in debug
This is due to lack of bounds checking in the debug fb implementation.
It is possible to fix it, but it would not be very useful anyway, since
there are only so many characters that could be shown at once using VGA
text mode and implementing scrolling would in the simplest way require a
PS/2 driver which make use of irq that is not available during early
boot.